chore: delete wasm.d.ts residual, rename agent-defaults, clean eval docs
Remove the Worker-era wasm.d.ts declaration, rename agent-defaults.ts to provider-defaults.ts for clarity, fix the eval playbook scope and comparison rules duplication, and clean up local eval workspace directories.

Made-with: Cursor
