--- src/wasm.d.ts
@@ -1,4 +1,0 @@
-declare module "*.wasm" {
-  const wasmModule: WebAssembly.Module
-  export default wasmModule
-}
